Move ChatViewModel to its own file and add tests.
